requirements.txt
asgiref==3.4.1
attrs==21.4.0
awsebcli==3.20.3
bcrypt==3.2.0
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==36.0.1
Django==3.2.11
django-cors-headers==3.11.0
django-extensions==3.1.5
django-sslserver==0.22
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.3
importlib-metadata==4.11.2
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.9.2
pathspec==0.9.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
pysqlite3==0.4.7
pysqlite3-binary==0.4.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.7.0
whitenoise
